Got able to work Add nd delete buttons

+++ Assignments 4.2/Student.cs
@@ -6,13 +6,13 @@ namespace Assignments_4._2
 {
     internal class Student
     {
-        public int StudentId { get; set; }
+        public double StudentId { get; set; }
         public string StudentName { get; set; }
         
         public Subject StudentSubject { get; set; }
 
         public double GPA { get; set; }
-        public Student(int studentId, string studentName, Subject studentSubject, double gpa)
+        public Student(double studentId, string studentName, Subject studentSubject, double gpa)
         {
             StudentId = studentId;
             StudentName = studentName;
